requirements.txt
click
loguru
ujson
requests
PyExecJS
rookiepy
pywebview
psutil
pyperclip